Restore maplayer for panorama_recent WFS
As it turns out the maplayer panorama_new doesn't play nice with WFS.
That needs to be addressed. In the meantime recent panoramas are
accesable in the old layer. Because the geo-view has been removed, the
query from that view has been pasted into the data-source. Styling of
dots has been copied from panorama_new layer.