Replace old Datapunt url with new Datapunt2 url
During migration all files were moved to new URLs, this commit changes
the location in the *.dbf files for LuFo tiles, so that the are being
picked up again by the mapserver.

This is just a textbased search and replace, this might not be enough,
in which case the shapefiles need to be generated again.