[DS-255] Added an attribute to use as parameter in URL
To select a specific tijdvak (there are four defined), the parameter tijdvak was added. It can be used in the URL like so: http://localhost:8383/maps/parkeervakken?tijdvak=00:00:00_06:59:00,07:00:00_17:00:00&version=1.3.0&service=WMS
The regime ('parkeervakken_tijdvakken_kiss_ride','parkeervakken_tijdvakken_laden_lossen',parkeervakken_tijdvakken_taxistandplaats') can be specified in the build in parameter layers.

It also possible to use call the WFS like this to get data for a specific tijdvak for a specific regime:
http://localhost:8383/cgi-bin/mapserv?map=/srv/mapserver/parkeervakken.map&service=WFS&version=1.1.0&request=GetFeature&typename=parkeervakken_tijdvakken_laden_lossen&featureID=parkeervakken_tijdvakken_laden_lossen.00:00:00_23:59:00&propertyName=tijdvak,begin_tijd,eind_tijd

Resolves: [DS-255]