Bouwstroompunt layers to energie WMS
Recently the bouwstroompunten is differentiated to bouwstroompunten for Markt, Evenement or general bouwstroompunt. There is also a sub division for categories of electric capacity. On data.amsterdam.nl the data on the map must be selectable. There will be the 3 base layers and within 4 sublayers. Like so:

Main:
   - Markt
   - Evenement
   - Bouwstroompunt

Sub:
   - tot 3 x 25 A
   - 3 x 25 A tot 3 x 40 A
   - 3 x 40 A tot 3 x 63 A
   - groter of gelijk aan 3 x 80 A

Because they each combination (main x sub) must be selectable (turn off/on), the combinations have there own layer in the WMS.

There is also an 'unkown' category added for data records without a capacity value.