Bug fixes for energie.map
This is a hotfix pending a redesign of the table to get rid of its messy
"capaciteit" column, for which we have to enumerate all possible values
because it's two numeric properties stashed in a single string column.

All filtering on "capaciteit" has been moved to SQL. This fixes a bug
and is less error-prone; the filtering had been broken by the removal of
parentheses in e2f01cb9. As a result, the categories got mixed up on the
map.

The layer "Evenement: tot 3 x 25 A" now matches the value '{3 X 16 A}',
which occurs four times in the refdb.

Finally, the abstract for the whole mapfile now matches the actual
content; energy use statistics were its original purpose, but got
removed in 717bca8707f8c5dc5b4f9aeb6a4766a69d336af6.

Closes #476.