AB#10367 Generate 403 Forbidden response for private maps if no JWT token
Because we are now exposing the JWT proxy externally, requests without a
JWT token are now passed on to the public mapserver instance.

However, this instance is missing the private maps, which would lead to
a 404. We want to generate a 403 for the private maps, so
a check is added for the incoming path and a 403 is added if needed.

One map (ondergrond) has a public and a private version (private has more
layers). We do not generate a 403 for this map. However, the mapserver
itself gives a sensible response about the layer not being available.