Simplify query for layer meetbout_zaksnelheid
Now two nested SELECTs instead of three. The left outer join is now an
inner join, which doesn't matter because (1) each meetbout has some
meting associated with it, and (2) if one didn't, it wouldn't match any
of the classes defined for this layer.

Also removed the manual version control comment.