Add BBOX test to CI/CD (#531)
We previously ran into MapServer/MapServer#6478 when trying to upgrade
our Ubuntu base image and added the file "epsg" as a hack to work around
it in 20d7dc720b53366424f42f2769b31c9477b172a0. This adds a regression
test for that upstream issue.

I have tested this with "COPY epsg /usr/share/proj" removed from the
Dockerfile, and that makes it fail.